Improve classes structure.
